requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.40
matplotlib>=3.7.0
PyYAML>=6.0.0
scipy>=1.10.0
pdfplumber>=0.11.0
